create setMatchMedia in assist.js
